rough draft 1 complete
Currently does Riemann sums fine. 
Only tested with a positive function, x^3.

The MC integration was getting huge errors because a (c,d) interval with negative values was used. Points below the curve but also below zero were added, where only the points below the curve and above zero should have been counted. Hence the ~50 error ([a,b] = [0,5] and [0,-10] was in the y-direction so 5*10 = 50).

Next step: fix the code to deal with negative functions for MC, and add functions/mechanisms that deal with improper input arguments.